# shared/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "StrategyCreate": {
        "additionalProperties": false,
        "description": "POST /api/strategies body.",
        "properties": {
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "name": {
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "plugin": {
            "anyOf": [
              {
                "maxLength": 64,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Plugin"
          },
          "slug": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Slug",
            "type": "string"
          },
          "status": {
            "default": "idea",
            "maxLength": 20,
            "title": "Status",
            "type": "string"
          },
          "tags": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tags"
          }
        },
        "required": [
          "name",
          "slug"
        ],
        "title": "StrategyCreate",
        "type": "object"
      },
...
      "StrategyRead": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "plugin": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Plugin"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "tags": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tags"
          },
          "versions": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/StrategyVersionRead"
            },
            "title": "Versions",
            "type": "array"
          }
        },
        "required": [
          "id",
          "name",
          "slug",
          "description",
          "status",
          "tags",
          "created_at"
        ],
        "title": "StrategyRead",
        "type": "object"
      },
...
      "StrategyUpdate": {
        "additionalProperties": false,
        "description": "PATCH /api/strategies/{id} body. Only fields present in the request\nare applied; omit a field to leave it untouched.",
        "properties": {
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "plugin": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Plugin"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "tags": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tags"
          }
        },
        "title": "StrategyUpdate",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Futures strategy research & backtesting API",
    "title": "BacktestStation",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/strategies": {
      "get": {
        "operationId": "list_strategies_api_strategies_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/StrategyRead"
                  },
                  "title": "Response List Strategies Api Strategies Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Strategies",
        "tags": [
          "strategies"
        ]
      },
      "post": {
        "operationId": "create_strategy_api_strategies_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StrategyCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StrategyRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Strategy",
        "tags": [
          "strategies"
        ]
      }
    },
...
    "/api/strategies/{strategy_id}": {
      "delete": {
        "description": "Delete a strategy \u2014 only allowed when it has zero versions.\n\nThis is the \"I created the wrong slug, clean it up\" path. It will NOT\ncascade to delete imported runs + trades + equity + metrics \u2014 if a\nstrategy has versions, we force the user to either delete each\nversion + its runs explicitly, or archive the strategy instead\n(PATCH status=\"archived\").\n\nReturns 409 if the strategy still has versions attached.",
        "operationId": "delete_strategy_api_strategies__strategy_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "strategy_id",
            "required": true,
            "schema": {
              "title": "Strategy Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Strategy",
        "tags": [
          "strategies"
        ]
      },
      "get": {
        "operationId": "get_strategy_api_strategies__strategy_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "strategy_id",
            "required": true,
            "schema": {
              "title": "Strategy Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StrategyRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Strategy",
        "tags": [
          "strategies"
        ]
      },
      "patch": {
        "operationId": "update_strategy_api_strategies__strategy_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "strategy_id",
            "required": true,
            "schema": {
              "title": "Strategy Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StrategyUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StrategyRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Strategy",
        "tags": [
          "strategies"
        ]
      }
    },
...
    "/api/strategies/{strategy_id}/runs": {
      "get": {
        "description": "All backtest runs across every version of this strategy.\n\nReplaces the frontend pattern of fetching /api/backtests and\nfiltering client-side \u2014 that approach is O(all runs in the system)\nwhich grows poorly.",
        "operationId": "list_strategy_runs_api_strategies__strategy_id__runs_get",
        "parameters": [
          {
            "in": "path",
            "name": "strategy_id",
            "required": true,
            "schema": {
              "title": "Strategy Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/BacktestRunRead"
                  },
                  "title": "Response List Strategy Runs Api Strategies  Strategy Id  Runs Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Strategy Runs",
        "tags": [
          "strategies"
        ]
      }
    },
    "/api/strategies/{strategy_id}/versions": {
      "post": {
        "operationId": "create_strategy_version_api_strategies__strategy_id__versions_post",
        "parameters": [
          {
            "in": "path",
            "name": "strategy_id",
            "required": true,
            "schema": {
              "title": "Strategy Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StrategyVersionCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StrategyVersionRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Strategy Version",
        "tags": [
          "strategies"
        ]
      }
    },
    "/api/strategy-versions/{version_id}": {
      "delete": {
        "description": "Delete a version \u2014 only allowed when it has zero attached runs.\n\nBefore this change, SQLAlchemy's cascade=\"all, delete-orphan\" on\nStrategyVersion.runs would silently nuke every imported run, trade,\nequity point, and metric with the version. Now the endpoint refuses\nwith 409 when runs exist and points the caller at the archive flow.",
        "operationId": "delete_strategy_version_api_strategy_versions__version_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "version_id",
            "required": true,
            "schema": {
              "title": "Version Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Strategy Version",
        "tags": [
          "strategies"
        ]
      },
      "patch": {
        "operationId": "update_strategy_version_api_strategy_versions__version_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "version_id",
            "required": true,
            "schema": {
              "title": "Version Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StrategyVersionUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StrategyVersionRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Strategy Version",
        "tags": [
          "strategies"
        ]
      }
    },
    "/api/strategy-versions/{version_id}/archive": {
      "patch": {
        "description": "Mark a version archived. Non-destructive \u2014 runs/trades/metrics untouched.",
        "operationId": "archive_strategy_version_api_strategy_versions__version_id__archive_patch",
        "parameters": [
          {
            "in": "path",
            "name": "version_id",
            "required": true,
            "schema": {
              "title": "Version Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StrategyVersionRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Archive Strategy Version",
        "tags": [
          "strategies"
        ]
      }
    },
    "/api/strategy-versions/{version_id}/baseline": {
      "patch": {
        "description": "Designate (or clear) the baseline run used by the Forward Drift Monitor.\n\nThe baseline is the run we expect live behavior to match. It must be\na non-live run (source=\"imported\" or \"engine\") \u2014 comparing live to\nitself is meaningless, so live runs are rejected with 422.\n\n`run_id=null` clears the baseline.",
        "operationId": "set_strategy_version_baseline_api_strategy_versions__version_id__baseline_patch",
        "parameters": [
          {
            "in": "path",
            "name": "version_id",
            "required": true,
            "schema": {
              "title": "Version Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StrategyVersionBaselineUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StrategyVersionRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Strategy Version Baseline",
        "tags": [
          "strategies"
        ]
      }
    },
    "/api/strategy-versions/{version_id}/unarchive": {
      "patch": {
        "operationId": "unarchive_strategy_version_api_strategy_versions__version_id__unarchive_patch",
        "parameters": [
          {
            "in": "path",
            "name": "version_id",
            "required": true,
            "schema": {
              "title": "Version Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StrategyVersionRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Unarchive Strategy Version",
        "tags": [
          "strategies"
        ]
      }
    },
...
  }
}